requirements.txt
# Core LanceDB dependencies
lancedb>=0.3.0
pyarrow>=14.0.0
pandas>=2.0.0
numpy>=1.24.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database & Caching
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
uvloop>=0.19.0
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.2.0
python-dateutil>=2.8.0
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0 